requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.5.14
decorator==5.1.1
distlib==0.3.6
faiss-cpu==1.7.4
filelock==3.12.0
Flask==2.3.2
Flask-Cors==4.0.0
fonttools==4.39.4
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
kiwisolver==1.4.4
langchain==0.0.251
langsmith==0.0.18
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib==3.7.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
platformdirs==3.5.1
preppy==4.2.2
protobuf==4.23.4
pyarrow==12.0.1
pydantic==1.10.12
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
PyMySQL==1.1.0
pyparsing==3.0.9
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.29.1
regex==2023.6.3
reportlab==4.0.4
requests==2.31.0
rich==13.4.2
rpds-py==0.8.10
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.19
streamlit==1.24.1
stripe==5.4.0
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.20.0
virtualenv==20.23.0
watchdog==3.0.0
Werkzeug==2.3.4
yarl==1.9.2
zipp==3.16.0
